pyproject.toml
[project]
name = "healthcare_ai_backend"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.2.1",
    "dateparser>=1.2.2",
    "livekit-agents[silero,turn-detector]~=1.2",
    "livekit-plugins-noise-cancellation~=0.2",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "python-dotenv>=1.1.1",
]
